requirements.txt
selenium>=3.141.0
Appium-Python-Client>=1.0.2
pytest>=6.1.2
pytest-html==2.1.1
pytest-html-reporter==0.2.3